ARTEMIS-5270 eliminate redundant casts

+++ tests/integration-tests/src/test/java/org/apache/activemq/artemis/tests/integration/journal/AsyncOpenCloseTest.java
@@ -119,7 +119,7 @@ public class AsyncOpenCloseTest extends ActiveMQTestBase {
             logger.debug("Closing");
             file.close(false, false);
             // even though the callback is blocked, the content of the file should already be good as written
-            validateFile(file, (byte) writtenByte);
+            validateFile(file, writtenByte);
             valve.countDown();
             Wait.assertEquals(0, submit::get, 5000, 10);
